Rename components editor <-> container for consistency
